Curses non-functional but now compiles.
