+++ bracket-terminal/src/hal/curses/font.rs
@@ -1,0 +1,16 @@
+use crate::Result;
+pub struct Font {
+    pub tile_size: (u32, u32),
+}
+
+impl Font {
+    pub fn load<S: ToString>(_filename: S, _tile_size: (u32, u32)) -> Font {
+        Font { tile_size: (1, 1) }
+    }
+
+    pub fn setup_gl_texture(&mut self, _gl: &crate::hal::BTermPlatform) -> Result<()> {
+        Ok(())
+    }
+
+    pub fn bind_texture(&self, _gl: &crate::hal::BTermPlatform) {}
+}
