+++ bracket-terminal/src/hal/curses/shader.rs
@@ -1,0 +1,1 @@
+pub struct Shader {}
